17) change amount from int to BibDecimal int method transfer, created new  HttpClient component

+++ src/main/java/com/example/bankservicedemo/controller/DataInsert.java
@@ -66,7 +66,7 @@ public class DataInsert {
         accountAlice.setUser(alice);
         accountService.save(accountAlice);
 
-        transactionService.transfer(accountBob, accountAlice, 100);
+        transactionService.transfer(accountBob, accountAlice, BigDecimal.valueOf(100));
         System.out.println("transaction successfully");
     }
 }
