+++ src/main/java/com/example/bankservicedemo/service/TransactionService.java
@@ -2,10 +2,11 @@ package com.example.bankservicedemo.service;
 
 import com.example.bankservicedemo.model.Account;
 import com.example.bankservicedemo.model.Transaction;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface TransactionService {
-    void transfer(Account fromAccount, Account toAccount, int amount);
+    void transfer(Account fromAccount, Account toAccount, BigDecimal amount);
 
     List<Transaction> getAllByAccount(int page, int size, Account account);
 }
